requirements.txt
aiohttp==3.9.3
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.2.1
tenacity==8.2.3
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0
lightgbm==4.3.0
xgboost==2.0.3
hmmlearn==0.3.0
torch==2.2.0
matplotlib==3.8.2
scikit-optimize==0.9.0
streamlit==1.31.0
plotly==5.18.0
requests==2.31.0
optuna==3.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
prometheus-client==0.19.0
cryptography==42.0.0
growwapi==0.0.1
